+++ packages/core/datetime-picker/examples/80-appearance.js
@@ -10,15 +10,22 @@ export default () => {
       <Label label="TimePicker - default no icon" />
       <TimePicker onChange={console.log} />
 
+      <Label label="TimePicker - subtle appearance" />
+      <TimePicker onChange={console.log} appearance="subtle" />
+
       <Label label="DatePicker - default" />
       <DatePicker onChange={console.log} />
       <Label label="DatePicker - hideIcon" />
       <DatePicker onChange={console.log} hideIcon />
+      <Label label="DatePicker - subtle appearance" />
+      <DatePicker onChange={console.log} appearance="subtle" />
 
       <Label label="DateTimePicker - default" />
       <DateTimePicker onChange={console.log} />
       <Label label="DateTimePicker - hideIcon" />
       <DateTimePicker onChange={console.log} hideIcon />
+      <Label label="DateTimePicker - subtle appearance" />
+      <DateTimePicker onChange={console.log} appearance="subtle" />
     </div>
   );
 };