+++ packages/core/breadcrumbs/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/breadcrumbs
 
+## 4.1.1
+
 ## 4.1.0
 - [minor] Added two new props: `itemsBeforeCollapse` and `itemsAfterCollapse`. With these, you can control how many items are displayed before and after the ellipsis in breadcrumbs' collapsed state. The default for both of these is one, meaning that the base behaviour is unchanged. [58bd739](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/58bd739)