+++ packages/core/field-text/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/field-text
 
+## 5.0.2
+
+## 5.0.1
+- [patch] Form developer preview [d8b2b03](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d8b2b03)
+- [patch] Form package developer preview release [9b28847](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9b28847)
+
 ## 5.0.0
 - [major] Bump to React 16.3. [4251858](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4251858)