+++ packages/core/tabs/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/tabs
 
+## 6.0.2
+
 ## 6.0.1
 - [patch] Update types [0073768](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/0073768)