+++ packages/core/toggle/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/toggle
 
+## 3.0.1
+
 ## 3.0.0
 - [major] Bump to React 16.3. [4251858](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4251858)