+++ packages/elements/pubsub/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/pubsub
 
+## 2.0.3
+
 ## 2.0.2
 - [patch] FS-1948 Stop pubsub client from flooding console with 403 [bb3d588](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/bb3d588)